+++ ts/editor/surround.ts
@@ -9,7 +9,8 @@ import type { SurroundFormat } from "../domlib/surround";
 import { boolMatcher, reformat, surround, unsurround } from "../domlib/surround";
 import { getRange, getSelection } from "../lib/cross-browser";
 import { registerPackage } from "../lib/runtime-require";
-import type { OnInsertCallback, Trigger } from "../sveltelib/input-manager";
+import type { OnInsertCallback } from "../sveltelib/input-manager";
+import type { Trigger } from "../sveltelib/trigger";
 import type { RichTextInputAPI } from "./rich-text-input";
 
 function isSurroundedInner(
